requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aiohttp==3.6.2            # via ray
async-timeout==3.0.1      # via aiohttp
attrs==19.3.0             # via aiohttp, jsonschema
beautifulsoup4==4.9.1     # via google
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via aiohttp, requests
click==7.1.2
colorama==0.4.3           # via ray
filelock==3.0.12          # via ray
future==0.18.2            # via torch
google==2.0.3             # via ray
grpcio==1.29.0            # via ray
idna==2.9                 # via requests, yarl
importlib-metadata==1.6.0  # via jsonschema
joblib==0.15.1            # via scikit-learn
jsonschema==3.2.0         # via ray
msgpack==0.6.2            # via ray
multidict==4.7.6          # via aiohttp, yarl
numpy==1.18.4
pillow==7.1.2             # via torchvision
protobuf==3.12.2          # via ray
py-spy==0.3.3             # via ray
pyrsistent==0.16.0        # via jsonschema
pyyaml==5.3.1             # via ray
ray==0.8.5
redis==3.4.1              # via ray
requests==2.23.0
scikit-learn==0.23.1
scipy==1.4.1              # via scikit-learn
six==1.15.0               # via grpcio, jsonschema, protobuf, pyrsistent
soupsieve==2.0.1          # via beautifulsoup4
threadpoolctl==2.1.0      # via scikit-learn
torch==1.5.0
torchvision==0.6.0
urllib3==1.25.9           # via requests
yarl==1.4.2               # via aiohttp
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools

# additional packages
python-dotenv
wandb
matplotlib
seaborn
kaleido

# install latex with the linux command: sudo apt install texlive-fonts-recommended texlive-fonts-extra dvipng